面向对象-类的属性(名词),方法(动词)介绍和使用

+++ src/com/company/object/Person.java
@@ -20,8 +20,8 @@ public class Person {
      */
     String name;
     int  age;
-    int  height;
-    int  weight;
+    double  height;
+    double  weight;
 
     /**
      * 吃饭的方法,行为
